Added an example for SAMD boards

+++ examples/double_logger/double_logger.ino
@@ -45,7 +45,7 @@ const char *sketchName = "logger_test.ino";
 const char *LoggerID = "SL099";
 const char *FileName = "doubleLoggerFile.csv";
 // Your logger's timezone.
-const int timeZone = -5;
+const int8_t timeZone = -5;
 // Create TWO new logger instances
 Logger logger1min;
 Logger logger5min;
@@ -57,13 +57,13 @@ Logger logger5min;
 #include <ProcessorStats.h>
 
 const long serialBaud = 57600;  // Baud rate for the primary serial port for debugging
-const int greenLED = 8;  // Pin for the green LED (-1 if unconnected)
-const int redLED = 9;  // Pin for the red LED (-1 if unconnected)
-const int buttonPin = 21;  // Pin for a button to use to enter debugging mode (-1 if unconnected)
-const int wakePin = A7;  // Interrupt/Alarm pin to wake from sleep
+const int8_t greenLED = 8;  // Pin for the green LED (-1 if unconnected)
+const int8_t redLED = 9;  // Pin for the red LED (-1 if unconnected)
+const int8_t buttonPin = 21;  // Pin for a button to use to enter debugging mode (-1 if unconnected)
+const int8_t wakePin = A7;  // Interrupt/Alarm pin to wake from sleep
 // Set the wake pin to -1 if you do not want the main processor to sleep.
 // In a SAMD system where you are using the built-in rtc, set wakePin to 1
-const int sdCardPin = 12;  // SD Card Chip Select/Slave Select Pin (must be defined!)
+const int8_t sdCardPin = 12;  // SD Card Chip Select/Slave Select Pin (must be defined!)
 
 const char *MFVersion = "v0.5";
 ProcessorStats mayfly(MFVersion) ;
@@ -73,9 +73,9 @@ ProcessorStats mayfly(MFVersion) ;
 //    Modem/Internet connection options
 // ==========================================================================
 HardwareSerial &ModemSerial = Serial1; // The serial port for the modem - software serial can also be used.
-const int modemSleepRqPin = 23;  // Modem SleepRq Pin (for sleep requests) (-1 if unconnected)
-const int modemStatusPin = 19;   // Modem Status Pin (indicates power status) (-1 if unconnected)
-const int modemVCCPin = -1;  // Modem power pin, if it can be turned on or off (-1 if unconnected)
+const int8_t modemSleepRqPin = 23;  // Modem SleepRq Pin (for sleep requests) (-1 if unconnected)
+const int8_t modemStatusPin = 19;   // Modem Status Pin (indicates power status) (-1 if unconnected)
+const int8_t modemVCCPin = -1;  // Modem power pin, if it can be turned on or off (-1 if unconnected)
 
 ModemSleepType ModemSleepMode = reverse;  // How the modem is put to sleep
 // Use "held" if the DTR pin is held HIGH to keep the modem awake, as with a Sodaq GPRSBee rev6.
@@ -102,7 +102,7 @@ MaximDS3231 ds3231(1);
 //    AOSong AM2315 Digital Humidity and Temperature Sensor
 // ==========================================================================
 #include <AOSongAM2315.h>
-const int I2CPower = 22;  // Pin to switch power on and off (-1 if unconnected)
+const int8_t I2CPower = 22;  // Pin to switch power on and off (-1 if unconnected)
 AOSongAM2315 am2315(I2CPower);
